Uses gosec GitHub action
Replaces the manual gosec installation with the official
GitHub action. This simplifies the CI configuration and ensures
that the latest version of gosec is used. It also removes the
need to manually download and install gosec, improving the
maintainability of the workflow.

Enables go modules by setting GO111MODULE=on to ensure consistent
builds across different environments.